rename the python script for long format ts
