+++ tests/python/unit/dku_timeseries/test_recipe_config_loading.py
@@ -1,6 +1,6 @@
 import pytest
 
-from recipe_config_loading import check_and_get_groupby_columns
+from long_format_config_loading import check_and_get_groupby_columns
 
 
 @pytest.fixture
